webapp/requirements.txt
affine==2.4.0
antlr4-python3-runtime==4.11.1
apache-age-python==0.0.7
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==4.2.4
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-environ==0.11.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
fastkml==0.12
fiona==1.9.5
GeoAlchemy2==0.14.3
geopandas==0.14.1
greenlet==3.0.3
lxml==4.9.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
psycopg2==2.9.9
pygeoif==0.7
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3
rasterio==1.3.9
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
soupsieve==2.4.1
SQLAlchemy==2.0.24
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
